Making zoneId optional for ZonedDateTime creation. **Breaking change**
Added a static defaultZoneId to ZonedDateTime that will be used if a
zoneId is not provided to a constructor or method. In most cases the
change is backwards compatible by making the zoneId an optional
parameter. However, that wasn't possible for the unnamed constructor,
`ZonedDateTime()`, so the zoneId parameter has been removed completely.
The original signature is preserved in the new
`ZonedDateTime.withZoneId()` constructor.